package: 외부에 의존하는 코드들을 io 패키지로 분리
