+++ src/main/java/domain/vo/AttendResult.java
@@ -1,11 +1,11 @@
-package dto;
+package domain.vo;
 
 import domain.AttendanceStatus;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public record AttendanceResponse(
+public record AttendResult(
         LocalDate attendDate,
         LocalTime attendTime,
         AttendanceStatus status
