+++ src/main/java/domain/vo/AttendanceRecordFindResults.java
@@ -1,4 +1,4 @@
-package vo;
+package domain.vo;
 
 import domain.ExpelWarning;
