+++ src/main/java/io/dto/ExpelWarningCrewResponse.java
@@ -1,4 +1,4 @@
-package dto;
+package io.dto;
 
 import domain.ExpelWarning;
