+++ src/main/java/io/reader/ConsoleReader.java
@@ -1,4 +1,4 @@
-package view.reader;
+package io.reader;
 
 import java.io.BufferedReader;
 import java.io.IOException;
