+++ src/main/java/io/view/InputView.java
@@ -1,8 +1,8 @@
-package view;
+package io.view;
 
-import dto.MenuSelect;
-import view.reader.Reader;
-import view.writer.Writer;
+import io.dto.MenuSelect;
+import io.reader.Reader;
+import io.writer.Writer;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
