+++ src/main/java/io/view/OutputView.java
@@ -1,10 +1,10 @@
-package view;
+package io.view;
 
 import domain.AttendanceStatus;
 import domain.ExpelWarning;
-import dto.AttendanceResponse;
-import dto.ExpelWarningCrewResponse;
-import view.writer.Writer;
+import io.dto.AttendanceResponse;
+import io.dto.ExpelWarningCrewResponse;
+import io.writer.Writer;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
