+++ src/main/java/io/writer/Writer.java
@@ -1,4 +1,4 @@
-package view.writer;
+package io.writer;
 
 public interface Writer {
