+++ src/main/resources/attendances.csv
@@ -1,82 +1,73 @@
 nickname,datetime
-쿠키,2024-12-02 13:30
-쿠키,2024-12-03 10:08
-쿠키,2024-12-04 09:30
-쿠키,2024-12-05 09:00
-쿠키,2024-12-06 11:00
-쿠키,2024-12-09 13:08
-쿠키,2024-12-10 10:05
-쿠키,2024-12-11 10:00
-쿠키,2024-12-12 10:00
-쿠키,2024-12-13 10:00
-쿠키,2024-12-16 10:00
-쿠키,2024-12-17 09:08
-쿠키,2024-12-18 10:15
-쿠키,2024-12-19 10:00
-쿠키,2024-12-20 10:05
-쿠키,2024-12-23 10:06
-쿠키,2024-12-24 10:00
-쿠키,2024-12-25 10:00
-쿠키,2024-12-26 10:00
-쿠키,2024-12-27 10:20
-쿠키,2024-02-01 11:20
-빙봉,2024-12-02 13:30
-빙봉,2024-12-03 10:08
-빙봉,2024-12-04 09:30
-빙봉,2024-12-05 09:00
-빙봉,2024-12-06 11:00
-빙봉,2024-12-09 13:08
-빙봉,2024-12-10 10:05
-빙봉,2024-12-11 10:00
-빙봉,2024-12-12 10:00
-빙봉,2024-12-13 10:00
-빙봉,2024-12-16 10:00
-빙봉,2024-12-17 09:08
-빙봉,2024-12-18 10:15
-빙봉,2024-12-19 10:00
-빙봉,2024-12-20 10:05
-빙봉,2024-12-23 10:06
-빙봉,2024-12-24 10:00
-빙봉,2024-12-25 10:00
-빙봉,2024-12-26 10:00
-빙봉,2024-12-27 10:20
-빙티,2024-12-02 13:02
-빙티,2024-12-03 10:05
-빙티,2024-12-04 10:04
-빙티,2024-12-05 09:35
-빙티,2024-12-06 09:59
-빙티,2024-12-09 12:49
-빙티,2024-12-10 10:02
-빙티,2024-12-11 10:01
-빙티,2024-12-12 10:03
-빙티,2024-12-13 09:50
-빙티,2024-12-16 12:50
-빙티,2024-12-17 10:01
-빙티,2024-12-18 10:02
-빙티,2024-12-19 09:59
-빙티,2024-12-20 09:20
-빙티,2024-12-23 12:59
-빙티,2024-12-24 10:01
-빙티,2024-12-25 10:03
-빙티,2024-12-26 10:04
-빙티,2024-12-27 10:02
-이든,2024-12-02 13:04
-이든,2024-12-03 10:04
-이든,2024-12-04 10:05
-이든,2024-12-05 10:04
-이든,2024-12-06 10:05
-이든,2024-12-09 13:06
-이든,2024-12-10 10:50
-이든,2024-12-11 10:20
-이든,2024-12-12 10:23
-이든,2024-12-13 10:01
-이든,2024-12-16 13:02
-이든,2024-12-17 09:50
-이든,2024-12-18 09:59
-이든,2024-12-19 10:02
-이든,2024-12-20 10:03
-이든,2024-12-23 13:05
-이든,2024-12-24 09:59
-이든,2024-12-25 09:45
-이든,2024-12-26 08:30
-이든,2024-12-27 10:30
+쿠키,2025-02-02 13:30
+쿠키,2025-02-03 10:08
+쿠키,2025-02-04 09:30
+쿠키,2025-02-05 09:00
+쿠키,2025-02-06 11:00
+쿠키,2025-02-09 13:08
+쿠키,2025-02-10 10:05
+쿠키,2025-02-11 10:00
+쿠키,2025-02-12 10:00
+쿠키,2025-02-13 10:00
+쿠키,2025-02-16 10:00
+쿠키,2025-02-17 09:08
+쿠키,2025-02-18 10:15
+쿠키,2025-02-19 10:00
+쿠키,2025-02-20 10:05
+쿠키,2025-02-23 10:06
+쿠키,2025-02-24 10:00
+쿠키,2025-02-25 10:00
+빙봉,2025-02-02 13:30
+빙봉,2025-02-03 10:08
+빙봉,2025-02-04 09:30
+빙봉,2025-02-05 09:00
+빙봉,2025-02-06 11:00
+빙봉,2025-02-09 13:08
+빙봉,2025-02-10 10:05
+빙봉,2025-02-11 10:00
+빙봉,2025-02-12 10:00
+빙봉,2025-02-13 10:00
+빙봉,2025-02-16 10:00
+빙봉,2025-02-17 09:08
+빙봉,2025-02-18 10:15
+빙봉,2025-02-19 10:00
+빙봉,2025-02-20 10:05
+빙봉,2025-02-23 10:06
+빙봉,2025-02-24 10:00
+빙봉,2025-02-25 10:00
+빙티,2025-02-02 13:02
+빙티,2025-02-03 10:05
+빙티,2025-02-04 10:04
+빙티,2025-02-05 09:35
+빙티,2025-02-06 09:59
+빙티,2025-02-09 12:49
+빙티,2025-02-10 10:02
+빙티,2025-02-11 10:01
+빙티,2025-02-12 10:03
+빙티,2025-02-13 09:50
+빙티,2025-02-16 12:50
+빙티,2025-02-17 10:01
+빙티,2025-02-18 10:02
+빙티,2025-02-19 09:59
+빙티,2025-02-20 09:20
+빙티,2025-02-23 12:59
+빙티,2025-02-24 10:01
+빙티,2025-02-25 10:03
+이든,2025-02-02 13:04
+이든,2025-02-03 10:04
+이든,2025-02-04 10:05
+이든,2025-02-05 10:04
+이든,2025-02-06 10:05
+이든,2025-02-09 13:06
+이든,2025-02-10 10:50
+이든,2025-02-11 10:20
+이든,2025-02-12 10:23
+이든,2025-02-13 10:01
+이든,2025-02-16 13:02
+이든,2025-02-17 09:50
+이든,2025-02-18 09:59
+이든,2025-02-19 10:02
+이든,2025-02-20 10:03
+이든,2025-02-23 13:05
+이든,2025-02-24 09:59
+이든,2025-02-25 09:45
