+++ src/test/java/domain/AttendanceBookTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import vo.AttendanceRecord;
-import vo.ExpelWarningResult;
+import domain.vo.AttendanceRecord;
+import domain.vo.ExpelWarningResult;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
