+++ src/test/java/domain/CrewAttendanceTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import vo.AttendanceRecord;
+import domain.vo.AttendanceRecord;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
